Moved getDashboardUrl to ServiceBrokerApiController
